requirements.txt
librosa==0.11.0
neucodec>=0.0.4
numpy==2.2.6
phonemizer==3.3.0
soundfile==0.13.1
transformers==4.56.1
resemble-perth==1.0.1
loguru
fastapi
uvicorn[standard]
python-multipart
jinja2
pyaudio
